Move non-event code away from event-specific partial type
